backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
paho-mqtt==1.6.1
pymongo==4.5.0
redis==5.0.0
python-dotenv==1.0.0
python-socketio==5.9.0
eventlet==0.33.3
python-engineio==4.7.1
dnspython==2.4.2
requests==2.31.0
jsonschema==4.19.0
pytest==7.4.2
pytest-flask==1.2.0
pytest-mock==3.11.1
authlib==1.2.1

# Testing dependencies
pytest-asyncio==0.21.1
pytest-cov==4.1.0
requests-mock==1.11.0

# Azure Functions
azure-functions==1.14.0
azure-functions-worker==1.6.0

# Notification services
vonage==3.14.0
twilio==8.12.0

# Email and webhooks
smtplib-ssl==1.0.4 authlib==1.6.1
